cloud_storage: add a object count limit to the cache
Previously we only limit on the basis of bytes: this made
sense when using large segments as disk space was generally
the limiting factor.  Now that we manage smaller sub-segment
chunks, we may hit metadata limits like the time it takes
to walk the tree to trim, and/or the size of the access
time tracker.

Fixes: https://github.com/redpanda-data/redpanda/issues/10726