cloud_storage: fix case where trimming isn't working
We already had a path for permitting cache reservations
to proceed after a failure to trim, if enough disk
space was available.  However, because the reserve
function always called trim_throttled, all reservations
would wait for up to 5 seconds when the system is
in this state.

Fix this by keeping a boolean flag to indicate if we
had hit this case on the last reservation, and use it
to selectively skip attempting to trim on the next
reservation.  The flag gets reset after each trim.