cloud_storage: account for reserved space when trimming
This fixes an edge case where if the reserved space was
>20% of the total cache allowance, trim() might fail to
trim anything because it sees that the current cache
size is already beneath the target.

This is only impactful in situations where the size of
reads in flight is of a similar order of magnitude of
the total cache size limit.  This hapens most in tests,
but as Redpanda runs on smaller nodes in the wild, it
will become realistic there too.