storage: only count remote topics in available bucket
When accounting for amount of data reclaimable by gc we will count bytes
that are subject to reclaim by the current local retention policy. for
cloud enabled topics we want to also estimate how much data in total is
reclaimable, and this is effectively how much data is present up to max
collectible offset.

prior to this commit we were counting this second bucket for non-cloud
enabled topics where max collectible offset is roughly the end of the
log in most cases. this led to misleading / incorrect stats about how
much data was available for reclaiming in gc process.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>